hotbody/back/hotbodyBack/requirements.txt
absl-py==0.10.0
appdirs==1.4.4
asgiref==3.2.10
astor==0.8.1
astroid==2.4.1
autopep8==1.5.3
beautifulsoup4==4.9.1
bleach==1.5.0
bootstrap4==0.1.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
coloredlogs==14.0
cycler==0.10.0
Cython==0.29.21
defusedxml==0.6.0
distlib==0.3.1
Django==3.1.1
django-allauth==0.42.0
django-bootstrap-pagination==1.7.1
django-bootstrap4==1.1.1
django-cors-headers==3.5.0
django-mysql==3.6.0
django-pandas==0.6.2
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
executor==23.1
fasteners==0.15
filelock==3.0.12
future==0.18.2
gast==0.4.0
grpcio==1.32.0
h5py==2.10.0
html5lib==0.9999999
humanfriendly==8.2
idna==2.10
importlib-metadata==1.7.0
inf==1.1.1
isort==4.3.21
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
linux-utils==0.7
Markdown==3.2.2
matplotlib==3.3.2
mccabe==0.6.1
mock==4.0.2
monotonic==1.5
numpy==1.19.2
oauthlib==3.1.0
opencv-python
pandas==1.1.2
Pillow==7.2.0
property-manager==3.0
protobuf==3.13.0
pycodestyle==2.6.0
PyJWT==1.7.1
pylint==2.5.2
pyparsing==2.4.7
pyreadline==2.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
scipy==1.5.2
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
toml==0.10.1
tqdm==4.50.0
typed-ast==1.4.1
urllib3==1.25.10
verboselogs==1.7
virtualenv==20.0.31
Werkzeug==1.0.1
wget==3.2
wrapt==1.12.1
zipp==3.1.0

# pytorch 
# pip install torch==1.6.0+cpu torchvision==0.7.0+cpu -f https://download.pytorch.org/whl/torch_stable.html
# pip install torch==1.5.1+cpu torchvision==0.6.1+cpu -f https://download.pytorch.org/whl/torch_stable.html
# pip install torch==1.5.0+cpu torchvision==0.6.0+cpu -f https://download.pytorch.org/whl/torch_stable.html